+++ src/multivariate/solvers/constrained/fminbox.jl
@@ -144,26 +144,28 @@ function optimize{T<:AbstractFloat,O<:Optimizer}(
     # initialization only makes use of the magnitude, we can fix this
     # by using the sum of the absolute values of the contributions
     # from each edge.
-    failidx = Array{Int,1}()
+    boundaryidx = Array{Int,1}()
     for i = 1:length(gbarrier)
         thisx = x[i]
         thisl = l[i]
         thisu = u[i]
 
-        if thisx <= thisl
+        if thisx == thisl
             thisx = 0.99*thisl+0.01*thisu
             x[i] = thisx
-            push!(failidx,i)
-        elseif thisx >= thisu
+            push!(boundaryidx,i)
+        elseif thisx == thisu
             thisx = 0.01*thisl+0.99*thisu
             x[i] = thisx
-            push!(failidx,i)
+            push!(boundaryidx,i)
+        elseif thisx < thisl || thisx > thisu
+            error("Initial position must be inside the box")
         end
 
         gbarrier[i] = (isfinite(thisl) ? one(T)/(thisx-thisl) : zero(T)) + (isfinite(thisu) ? one(T)/(thisu-thisx) : zero(T))
     end
-    if length(failidx) > 0
-        warn("Initial position must be inside the box. Moving initial condition elements with index:\n$failidx")
+    if length(boundaryidx) > 0
+        warn("Initial position cannot be on the boundary of the box. Moving elements to the interior.\nElement indices affected: $boundaryidx")
     end
     df.g!(gfunc, x)
     mu = isnan(mu0) ? initial_mu(gfunc, gbarrier; mu0factor=mufactor) : mu0