+++ NEWS.md
@@ -1,5 +1,5 @@
 # Optim v1.0.0
-* Fminbox: If an initial guess is not in the interior of the box, the guess is moved inside the box and a warning is produced, as opposed to exting with an error.
+* Fminbox: If an initial guess is on the boundary of the box, the guess is moved inside the box and a warning is produced, as opposed to crashing with an error.
 * Significant changes to the Non-, Once-, and TwiceDifferentiable setup; these now hold temporaries relevant to the evaluation of objectives, gradients, and Hessians. They also hold f-, g-, and h_calls counters
 * Refactor tests
 * Drop v0.4 support